requirements.txt
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Babel==1.0.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
urllib3==1.25.9
Werkzeug==1.0.1
